refactor: 파일명 수정 및 final 키워드 누락 수정
