+++ src/test/java/nextstep/repository/ConsoleReservationRepositoryTest.java
@@ -1,6 +1,6 @@
 package nextstep.repository;
 
-import nextstep.Theme;
+import nextstep.domain.theme.Theme;
 import nextstep.domain.reservation.Reservation;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -12,8 +12,8 @@ import java.time.LocalTime;
 
 import static org.assertj.core.api.Assertions.*;
 @SpringBootTest
-public class ConsoleReservationRepoTest {
-    private final ReservationRepo consoleReservationRepo = new ConsoleReservationRepo();
+public class ConsoleReservationRepositoryTest {
+    private final ReservationRepo consoleReservationRepo = new ConsoleReservationRepository();
 
     @DisplayName("reservation test")
     @Test
@@ -25,20 +25,20 @@ public class ConsoleReservationRepoTest {
                 new Theme("워너고홈", "병맛 어드벤처 회사 코믹물", 29_000)
         );
 
+        // 추가
         long id = consoleReservationRepo.add(newReservation);
         Reservation reservation = consoleReservationRepo.findById(id).orElseThrow();
-
         assertThat(reservation).isEqualTo(newReservation);
 
+        // 예약 시간 중복
         int countSameDateAndTime = consoleReservationRepo.countByDateAndTime(
                 Date.valueOf(reservation.getDate()),
                 Time.valueOf(reservation.getTime()));
         assertThat(countSameDateAndTime > 0).isTrue();
 
+        // 삭제
         consoleReservationRepo.delete(id);
-
         Reservation reservation2 = consoleReservationRepo.findById(id).orElse(null);
-
         assertThat(reservation2).isNull();
     }
 }
